Updating config.yaml to indicate that After6 Services is providing support
We are replacing references to Byrne Reese, Melody, and Open Melody with After6 Services and Movable Type 5/6. The reasons are because Byrne is no longer maintaining Config Assistant, Open Melody Software Group is defunct, and After6 Services is maintaining Config Assistant going for Movable Type 5 and 6 going forward.

We maintain references to Byrne Reese, Jay Allen, and Dan Wolfgang in README.md.